Remove old scope support
