+++ packages/cursorless-org-docs/src/docs/user/languages/components/Language.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import { ScopeSupport } from "./ScopeSupport";
-import { ScopeVisualizer } from "./ScopeVisualizer";
 
 interface Props {
   languageId: string;
@@ -9,10 +8,6 @@ interface Props {
 export function Language({ languageId }: Props) {
   return (
     <>
-      <h2>Scopes</h2>
-
-      <ScopeVisualizer languageId={languageId} />
-
       <ScopeSupport languageId={languageId} />
     </>
   );
