+++ packages/cursorless-org-docs/src/docs/user/languages/components/ScopeSupport.css
